- Se corrigieron infinitos problemas en FiniteBestStrip - Se arreglo el main para hacer lectura de archivos grandes - Se hicieron funciones search4strips y search4others - Se cambion merge_strip a forma iterativa sin iteradores de vectores - Se comentó por el momento algunas partes de printpacking

+++ HFirstBest.cpp
@@ -11,7 +11,7 @@ int HFirstBest(int Tbin, Packing items, int Hbin, int Wbin){
   long combs = comb((n-1),k);
   cout<<"combs:"<<combs<<"\n";
   if (combs > 100000)
-    combs = combs*(0.4);
+    combs = combs* (long) (0.4);
   
   //Variables temporales de la heuristica
   Packing pack;
